Home

ODBC.jl

The ODBC.jl package provides high-level julia functionality over the low-level ODBC API. In particular, the package allows making connections with any database that has a valid ODBC driver, sending SQL queries to those databases, and streaming the results into a variety of data sinks.

ODBC.jl

ODBC administrative functions

DBMS Connections

Query execution and result handling

ODBC administrative functions

ODBC.drivers — Function.

ODBC.drivers() -> Dict

List installed ODBC drivers. The primary config location for installed drivers on non-windows platforms is realpath(joinpath(dirname(pathof(ODBC)), "../config/odbcinst.ini")), i.e. an ODBC.jl-managed location. Other system/user locations may also be checked (and are used by default on windows) by the underlying ODBC driver manager, but for the most consistent results, aim to allow ODBC.jl to manage installed drivers/datasources via ODBC.addriver, ODBC.removedriver, etc.

On windows, ODBC.jl uses the system-wide configurations for drivers and datasources. Drivers and datasources can still be added via ODBC.adddriver/ODBC.removdriver and ODBC.adddsn/ODBC.removedsn, but you must have administrator privileges in the Julia session. This is accomplished easiest by pressing CTRL then right-clicking on the terminal/Julia application and choosing "Run as administrator".

source

ODBC.dsns — Function.

ODBC.dsns() -> Dict

List installed ODBC datasources. The primary config location for installed datasources on non-windows platforms is realpath(joinpath(dirname(pathof(ODBC)), "../config/odbc.ini")), i.e. an ODBC.jl-managed location. Other system/user locations may also be checked (and are by default on windows) by the underlying ODBC driver manager, but for the most consistent results, aim to allow ODBC.jl to manage installed drivers/datasources via ODBC.adddsn, ODBC.removedsn, etc.

On windows, ODBC.jl uses the system-wide configurations for drivers and datasources. Drivers and datasources can still be added via ODBC.adddriver/ODBC.removdriver and ODBC.adddsn/ODBC.removedsn, but you must have administrator privileges in the Julia session. This is accomplished easiest by pressing CTRL then right-clicking on the terminal/Julia application and choosing "Run as administrator".

source

ODBC.adddriver — Function.

ODBC.adddriver(name, libpath; kw...)

Install a new ODBC driver. name is a user-provided "friendly" name to identify the driver. libpath is the absolute path to the ODBC driver shared library. Other key-value driver properties can be provided by the kw... keyword arguments.

This method is provided to try and provide the simplest/easiest/most consistent setup experience for installing a new driver. Editing configuration files by hand is error-prone and it's easy to miss adding something that is required.

While ODBC.jl supports all 3 major ODBC driver managers (unixODBC, iODBC, and odbc32), be aware that most DBMS ODBC driver libraries are built against only one of the 3 and can lead to compatibility issues if a different driver manager is used. This is mainly an issue for driver libraries built against iODBC and then tried to use with unixODBC or vice-versa.

On windows, ODBC.jl uses the system-wide configurations for drivers and datasources. Drivers and datasources can still be added via ODBC.adddriver/ODBC.removdriver and ODBC.adddsn/ODBC.removedsn, but you must have administrator privileges in the Julia session. This is accomplished easiest by pressing CTRL then right-clicking on the terminal/Julia application and choosing "Run as administrator".

source

ODBC.removedriver — Function.

ODBC.removedriver(name; removedsns::Bool=true)

Remove an installed ODBC driver by name (as returned from ODBC.drivers()). removedsns=true also removes any datasources that were specified to use the driver.

On windows, ODBC.jl uses the system-wide configurations for drivers and datasources. Drivers and datasources can still be added via ODBC.adddriver/ODBC.removdriver and ODBC.adddsn/ODBC.removedsn, but you must have administrator privileges in the Julia session. This is accomplished easiest by pressing CTRL then right-clicking on the terminal/Julia application and choosing "Run as administrator".

source

ODBC.adddsn — Function.

ODBC.adddsn(name, driver; kw...)

Install a new ODBC datasource. name is a user-provided "friendly" name to identify the datasource (dsn). driver is the "friendly" driver name that should be used to connect to the datasource (valid driver options can be seen from ODBC.drivers()). Additional connection key-value properties can be provided by the kw... keyword arguments.

Datasources can be connected by calling DBInterface.connect(ODBC.Connection, dsn, user, pwd), where dsn is the friendly datasource name, user is the username, and pwd is the password.

An alternative approach to installing datasources is to generate a valid "connection string" that includes all connection properties in a single string passed to DBInterface.connect. www.connectionstrings.com is a convenient resource that provides connection string templates for various database systems.

On windows, ODBC.jl uses the system-wide configurations for drivers and datasources. Drivers and datasources can still be added via ODBC.adddriver/ODBC.removdriver and ODBC.adddsn/ODBC.removedsn, but you must have administrator privileges in the Julia session. This is accomplished easiest by pressing CTRL then right-clicking on the terminal/Julia application and choosing "Run as administrator".

source

ODBC.removedsn — Function.

ODBC.removedsn(name)

Remove an installed datasource by name (as returned from ODBC.dsns()).

On windows, ODBC.jl uses the system-wide configurations for drivers and datasources. Drivers and datasources can still be added via ODBC.adddriver/ODBC.removdriver and ODBC.adddsn/ODBC.removedsn, but you must have administrator privileges in the Julia session. This is accomplished easiest by pressing CTRL then right-clicking on the terminal/Julia application and choosing "Run as administrator".

source

ODBC.setdebug — Function.

ODBC.setdebug(debug::Bool=true, tracefile::String=joinpath(tempdir(), "odbc.log"))

Turn on ODBC library call tracing. This prints debug information to tracefile upon every entry and exit from calls to the underlying ODBC library (unixODBC, iODBC, or odbc32). Debugging can be turned off by passing false.

Note that setting tracing on/off requires resetting the ODBC environment, which means any open statements/connections will be closed/invalid.

Also note that due to the persistent nature of ODBC config, setting tracing will persist acrosss Julia sessions, i.e. if you turn tracing on, then quit julia and start again tracing will still be on, and will stay on until explicitly turned off.

The iODBC driver manager supports passing stderr as the tracefile, which will print all tracing information into the julia session/repl.

source

DBMS Connections

DBInterface.connect — Function.

DBInterface.connect(dsn_or_connectionstring, user, password; connectionstring::Bool=false)

Construct a Connection type by connecting to a valid ODBC Connection or by specifying a datasource name or valid connection string. Takes optional 2nd and 3rd arguments for named datasources username and password, respectively. 1st argument dsn can be either the name of a pre-defined ODBC Connection or a valid connection string. If passing a connection string, the connectionstring=true keyword argument must also be passed. The user and pwd arguments are ignored if connectionstring=true. A great resource for building valid connection strings is http://www.connectionstrings.com/.

source

ODBC.Connection — Type.

ODBC.Connection

A Connection represents an established ODBC connection. It is passed to most other ODBC methods as a first argument See DBInterface.execute for information on making ODBC connections to database systems.

source

DBInterface.close! — Function.

DBInterface.close!(conn)

Close an open connection. In general, statements and open cursors will not be valid once a connection has been closed.

source

DBInterface.close!(stmt)

Close a prepared statement. Further parameter binding or execution will not be valid.

source

Query execution and result handling

DBInterface.prepare — Function.

DBInterface.prepare(conn, sql) -> ODBC.Statement

Prepare a query string, optionally including parameters to bind upon execution (with ? markers). Please refer to individual dbms documentation for the exact level of parameter binding support.

The returned prepared statement can then be passed to DBInterface.execute(stmt, params) with params that will be bound before execution. This allows preparing the statement once, and re-using it many times with different parameters (or the same) efficiently.

source

DBInterface.execute — Function.

DBInterface.execute(stmt, params=(); kw...) -> ODBC.Cursor

Execute a prepare statement, binding any parameters beforehand. Returns a Cursor object, even if the statement is not resultset-producing (cursor will have zero rows and/or columns). The Cursor object satisfies the Tables.jl interface as a source, so any valid sink can be used for inspecting results (a list of integrations is maintained here). Supported keyword arguments include iterate_rows::Bool for forcing row iteration of the resultset, and debug::Bool for printing additional debug information during the query/result process.

source

DBInterface.execute(conn, sql, params=(); kw...) -> ODBC.Cursor

Send a query directly to connection for execution. Returns a Cursor object, even if the statement is not resultset-producing (cursor will have zero rows and/or columns). The Cursor object satisfies the Tables.jl interface as a source, so any valid sink can be used for inspecting results (a list of integrations is maintained here). Supported keyword arguments include iterate_rows::Bool for forcing row iteration of the resultset, and debug::Bool for printing additional debug information during the query/result process.

This is an alternative execution path to DBInterface.execute with a prepared statement. This method is faster/less overhead for one-time executions, but prepared statements will have more benefit for repeated executions (even with different parameters).

source

Missing docstring.

Missing docstring for DBInterface.executemultiple. Check Documenter's build log for details.